==> scripts/0_setup.R <==

# Load packages

library("readxl")
library("dplyr")
library("tidyr")
library("readr")
library("ggplot2")
library("forcats")
library("purrr")
library("stringr")
library("janitor")
library("glue")

# Ensure that raw data is saved to "data/raw/fy26-30-adopted-capital-open-data.xlsx"
# From https://data.boston.gov/dataset/capital-budget/resource/c62d666e-27ea-4c03-9cb1-d3a81a1fb641
